requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
tzdata==2025.1
pandas>=1.3.0
numpy>=1.20.0
requests>=2.26.0
smartapi-python==1.5.5
python-dotenv>=0.19.0
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
ta>=0.10.1
matplotlib>=3.4.0
